Minor refactoring in the code
