+++ src/main/java/dev/farhan/movieist/movies/controller/MovieController.java
@@ -1,9 +1,13 @@
-package dev.farhan.movieist.movies;
+package dev.farhan.movieist.movies.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import dev.farhan.movieist.movies.model.Movie;
+import dev.farhan.movieist.movies.service.MovieService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 import java.util.Optional;
@@ -12,8 +16,11 @@ import java.util.Optional;
 @RequestMapping("/api/v1/movies")
 public class MovieController {
 
-    @Autowired
-    private MovieService service;
+    private final MovieService service;
+
+    public MovieController(MovieService service) {
+        this.service = service;
+    }
 
     @GetMapping
     public ResponseEntity<List<Movie>> getMovies() {
