+++ src/main/java/dev/farhan/movieist/movies/controller/ReviewController.java
@@ -1,0 +1,29 @@
+package dev.farhan.movieist.movies.controller;
+
+import dev.farhan.movieist.movies.model.Review;
+import dev.farhan.movieist.movies.service.ReviewService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Map;
+
+@RestController
+@RequestMapping("/api/v1/reviews")
+public class ReviewController {
+
+    private final ReviewService service;
+
+    public ReviewController(ReviewService service) {
+        this.service = service;
+    }
+
+    @PostMapping()
+    public ResponseEntity<Review> createReview(@RequestBody Map<String, String> payload) {
+
+        return new ResponseEntity<>(service.createReview(payload.get("reviewBody"), payload.get("imdbId")), HttpStatus.OK);
+    }
+}
