+++ src/main/java/dev/farhan/movieist/movies/repository/MovieRepository.java
@@ -1,5 +1,6 @@
-package dev.farhan.movieist.movies;
+package dev.farhan.movieist.movies.repository;
 
+import dev.farhan.movieist.movies.model.Movie;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
