+++ src/main/java/dev/farhan/movieist/movies/service/ReviewService.java
@@ -1,21 +1,25 @@
-package dev.farhan.movieist.movies;
+package dev.farhan.movieist.movies.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import dev.farhan.movieist.movies.model.Movie;
+import dev.farhan.movieist.movies.model.Review;
+import dev.farhan.movieist.movies.repository.ReviewRepository;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
-import java.util.Optional;
 
 @Service
 public class ReviewService {
-    @Autowired
-    private ReviewRepository repository;
 
-    @Autowired
-    private MongoTemplate mongoTemplate;
+    private final ReviewRepository repository;
+    private final MongoTemplate mongoTemplate;
+
+    public ReviewService(ReviewRepository repository, MongoTemplate mongoTemplate) {
+        this.repository = repository;
+        this.mongoTemplate = mongoTemplate;
+    }
 
     public Review createReview(String reviewBody, String imdbId) {
         Review review = repository.insert(new Review(reviewBody, LocalDateTime.now(), LocalDateTime.now()));
