--- src/main/resources/.env.example
@@ -1,4 +1,0 @@
-MONGO_DATABASE=
-MONGO_USER=
-MONGO_PASSWORD=
-MONGO_CLUSTER=
